refactor(kmeans): reorg
