--- docs/algorithms/clustering/diana.md
@@ -1,3 +1,0 @@
-# DIANA
-
-::: toyml.clustering.diana
