+++ docs/algorithms/clustering/kmeans.md
@@ -1,3 +1,5 @@
 # Kmeans
 
-::: toyml.clustering.kmeans_clustering
+::: toyml.clustering.kmeans.simple
+::: toyml.clustering.kmeans.plus
+::: toyml.clustering.kmeans.bisect
