+++ toyml/clustering/kmeans/__init__.py
@@ -1,0 +1,5 @@
+from .bisect import BisectingKmeans
+from .plus import KmeansPlus
+from .simple import Kmeans
+
+__all__ = ["Kmeans", "BisectingKmeans", "KmeansPlus"]
